Quit, fixed "pid"
Added a way to safely quit the program. Pressing the gray button ends
the main loop and starts the shutdown process, turning everything off.

Also updated the "pid" for the lift motors. They now go to within a
range of their targets, as opposed to trying for the actual targets
themselves. Made the code more compact, too. Put in some dummy encoder
targets.

Changed the encoder test program for use in determining the targets for
all our things.

Started putting in some encoder and servo targets